Remove NamespaceIdMap
